refactor(lib): seperate link list and bubblesort into header and source files
