+++ link_list/CMakeLists.txt
@@ -1,0 +1,2 @@
+project(link_list)
+add_library(link_list link_list.c)
