+++ sorting/bubble_sort/CMakeLists.txt
@@ -1,0 +1,2 @@
+project(bubble_sort)
+add_library(bubble_sort bubble_sort.c)
